+++ src/main/scala/com/github/unchama/seichiassist/itemconversionstorage/TrashBox.scala
@@ -1,16 +1,16 @@
-package com.github.unchama.seichiassist.processingstorage
+package com.github.unchama.seichiassist.itemconversionstorage
 
 import cats.effect.IO
+import com.github.unchama.itemconversionstorage.ItemConversionStorage
 import com.github.unchama.menuinventory.syntax.IntInventorySizeOps
 import com.github.unchama.menuinventory.{MenuFrame, MenuSlotLayout}
-import com.github.unchama.processingstorage.ProcessingStorage
 import org.bukkit.entity.Player
-import org.bukkit.inventory.{InventoryView, ItemStack}
+import org.bukkit.inventory.ItemStack
 
 /**
  * ゴミ箱
  */
-object TrashBox extends ProcessingStorage {
+object TrashBox extends ItemConversionStorage {
   override type Environment = ()
   override val frame: MenuFrame = MenuFrame(4.chestRows, "ゴミ箱")
